Add MLIR dump utility that mirrors dump_graph utils
The MLIR textual IR form is (IMHO) easier to read than the protobuf output, so add a convenience method to dump the IR form that mirrors the protobuf dumper form. Also add option to enable running standard pipeline (which often produces prettier form as a plus but could obscure state [e.g., it could show cleaner graph than what the rest of the execution will see on the current stack]).

This is just the basic dumper, more effort may be needed to make it work well & printing syntax could be enhanced (for which tests need to be updated).

PiperOrigin-RevId: 305763898
Change-Id: Ib2eb77d92c74f44ec159cc232ca1e67fcad0209d