Improve @tf.function retracing detection coverage.
Previously, the frequent retracing detection logic was per @tf.function instance,
So it couldn't detect cases that @tf.function instance is frequently recreated.

For example, the following didn't raise a frequent retracing warning.
```Python
for i in range(10):
  @tf.function
  def test(x):
    return x
  test(i)
```

Move the retracing detection to the global scope and use `__code__` (or Python function instance if not available) as a key to enable frequent retracing detection across @tf.function lifetimes.

PiperOrigin-RevId: 306238536
Change-Id: I045fbbb51625c34491be970d18088e96d5ff47dc