Add type op interface to tf.const
Enables using tf.const op via type inference op interface and generating
generic build method. Currently the additional build method there is more
generic (e.g., it does some wrapping of non-elementsattr's that is not
supported by the op) and so is retained.

This also uses the C++ trait rather than any generated one. Using op with very simple transfer function that is easy to test (and it also showed a bug in a test :))

PiperOrigin-RevId: 307870211
Change-Id: I328c7f002fea3445027f7fbbd9bccdfdf2da3940