Error out from ScopedAllocatorOptimizer when input shape is not fully known.
ScopedAllocatorOptimizer would invoke `TensorShape::IsValid` on the input.  It
was assumed this would be true when the shape is fully known, but
`TensorShape::IsValid` can be true when the number of dimensions in the shape
is unknown as well.

This change fixes the shape check so that we now exclude ops which have unknown
rank from this optimization.

This change also adds a unit test for ScopedAllocatorOptimizer + XLA JIT
compilation, which unearthed this bug.

PiperOrigin-RevId: 308123606
Change-Id: I185e1e890d1fc0533635d5325d22f71b3a8480b4