[XLA:CPU] Reconsider alignment heuristics
We were trying to model tcmalloc and glibc internals based on completely
outdated heuristics. operator new will return any alignment based on the type,
so using e.g. memory backed by a std::vector was never an option here. The
heuristic was also known to produce annoying performance cliffs (e.g. with a
convolution with a tiny filter).

Just unify all the places we hardcoded alignment to use
cpu_function_runtime::kMinAlign and make it use 16, which is the sweet spot for
vectorization on most platforms. Assert that alignment is sufficient when
giving buffers to tfcompile'd code. This also matches the default alignment
in Eigen, so we stay compatible. Raising it would break things like
Eigen::aligned_allocator.

We still align buffers in tfcompile to 64 bytes for optimal performance. I
tried lowering this to 16, but that triggered some benchmarks to slow down.

PiperOrigin-RevId: 308222775
Change-Id: Ib308f4d687a7b27ae5016365080ad957fe8584bc