Remove the necessity of acquiring context_update_mu_ in cluster update.
After this change, we no longer guarantee that an enqueue request does not overlap with a cluster update request. It means that a function execution can see inconsistent cluster views (e.g., some part using old server def and some with new one). Function executed in parallel with or shortly after cluster update requests can potentially fail on connection errors due to this inconsistency if the functions span remote targets that are being updated. To prevent this race condition from happening, the client needs to coordinate and avoid concurrent executing remote functions and updating the cluster.

PiperOrigin-RevId: 308342265
Change-Id: I57fd5f3ec9acd1da906d2718d7b61398156bb4f4