Refactor the unified C API implementation in a private header
The idea is to prepare to split the various implementation of the API
by exposing the API through abstract base class in this private headers
with specific implementations for Graph/Eager/MLIR in different files.
The C API should be able to be implemented as a binding over the API
exposed by the abstract base classes and independent of the actual
implementations.

PiperOrigin-RevId: 308387700
Change-Id: I0610d4e9b8067a14d2b31e3c4777d8ed783987c6