Refactor AbstractionFunction implementation to make it abstract on the C++ side as well
This is an interesting exercise as it is surfacing some coupling between eager
and graph. This isn't surprising as this coupling existing in the existing runtime,
but as we're trying to untangle every pieces such refactoring is forcing to expose
the coupling by introducing a new API to extract a TF_Function out of the
AbstractFunction.

PiperOrigin-RevId: 308423040
Change-Id: I49f4fed4cb6d1c82d4b7af33cb9d5eac7edcf703