Allow functions with raw Send & Recv to be inlined
We allow functions that have Send & Recv to be inlined. This causes them
to use the same rendezvous as the outer function, allowing functions to
communicate with each other.

Further, We all Recv to be run asynchronously but ensure Send is added
added as a control output so that is doesn't get pruned.

PiperOrigin-RevId: 309314078
Change-Id: I485bd3486e1ff40ed1aee2b157780ccdbac681df