[XLA]Add HLO infrastructure to support dynamic op lowering.
This CL gives additional configuration to dynamic padder, which tells it what ops can have accept dynamic tensors.

If an op requires dynamic tensor as input, and a static tensor is presented, a "SliceToDynamic" will be inserted.

If an op requires static tensor as input and a dynamic tensor is presented, a
"PadToStatic" op will be inserted.

If an op requires static tensor and the tensor is already static, dynamic padder will rewrite the op to make it produce the same result as if the tensor is dynamic (this is what we already have today).

PiperOrigin-RevId: 309326119
Change-Id: I5376674d6acf9905af1b7e09b127811b57517e97